Update reference binaries for min/max reduce script.
